Переработка утилит
